[breaking] Remove unnecessary re-export.
