+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "npm run --silent tests-only",
     "tests-only": "mocha test/*.js --reporter spec",
-    "docs": "docco -o docs/ lib/polyglot.js"
+    "docs": "docco -o docs/ index.js"
   },
   "repository": {
     "type": "git",
